Add envs to check whether to use free ChatGPT; Add feature to continue processing exhibitors

+++ core/__init__.py
@@ -1,4 +1,5 @@
 import csv
+import os
 from typing import Dict, List
 
 FILE_NAME = 'exhibitors.csv'
@@ -9,6 +10,8 @@ def prepare_writer() -> None:
     """
     Prepare file to write data
     """
+    if os.path.isfile(FILE_NAME):
+        return
     with open(FILE_NAME, mode='w') as exhibitors_file:
         fieldnames = FIELD_NAMES
         writer = csv.DictWriter(exhibitors_file, fieldnames=fieldnames)
@@ -25,3 +28,16 @@ def write_data(
         fieldnames = FIELD_NAMES
         writer = csv.DictWriter(exhibitors_file, fieldnames=fieldnames)
         writer.writerow(data)
+
+
+def get_companies() -> List[str]:
+    """
+    Prepare file to write data
+    """
+    companies = []
+    with open(FILE_NAME, mode='r') as exhibitors_file:
+        fieldnames = FIELD_NAMES
+        reader = csv.DictReader(exhibitors_file, fieldnames=fieldnames)
+        for row in reader:
+            companies.append(row['company'])
+    return companies
